refactor: 정적 데이터 관련 리팩토링
