+++ src/mappers/index.ts
@@ -1,4 +1,5 @@
 export * from './promptMapper';
 export * from './promptFormMapper';
+export * from './metaMapper';
 
 export * from './mapper.types';
