+++ src/services/index.ts
@@ -1,1 +1,4 @@
-export * from '@/services/prompt';
+export * from './prompt';
+
+export * from './common/meta';
+export * from './common/meta.types';
